notif [nfc]: Introduce NotificationOpenService
And move the notification navigation data parsing utilities to
the new class.
